requirements.txt
Flask
gspread
google-auth
pandas
git+https://github.com/bananaya/TaiwanLotteryCrawler.git@main
requests
urllib3
